refactor(assistant-service): Extracted endpoint routing from main.py
